:recycle: :construction: refactoring category drawer
